Extract logic for handling clicks away in dropdowns to its own component, bring dropdown menu to front
