+++ ui/src/SpaceDashboard/SpaceDashboardTile.scss
@@ -109,6 +109,7 @@
 
 
 .ellipsisDropdownContainer {
+  z-index: 1;
   position: absolute;
   margin-top: 8px;
   background-color: white;
